+++ commercetools/resource_subscription_test.go
@@ -34,6 +34,11 @@ func TestValidateDestination(t *testing.T) {
 			"project_id": "<project_id>",
 			"topic":      "<topic>",
 		},
+		{
+			"type":       "event_bridge",
+			"region":     "<region>",
+			"account_id": "<account_id>",
+		},
 	}
 	for _, validDestination := range validDestinations {
 		_, errs := validateDestination(validDestination, "destination")
@@ -57,6 +62,10 @@ func TestValidateDestination(t *testing.T) {
 		{
 			"type":  "google_pubsub",
 			"topic": "<topic>",
+		},
+		{
+			"type":  "event_bridge",
+			"topic": "<region>",
 		},
 	}
 	for _, validDestination := range invalidDestinations {